+++ contrib/log_correlation/log4j2/src/test/java/io/opencensus/contrib/logcorrelation/log4j2/AbstractOpenCensusLog4jLogCorrelationTest.java
@@ -45,8 +45,7 @@ abstract class AbstractOpenCensusLog4jLogCorrelationTest {
   private static final Tracer tracer = Tracing.getTracer();
 
   static final String TEST_PATTERN =
-      "traceId=%X{opencensusTraceId} spanId=%X{opencensusSpanId} "
-          + "sampled=%X{opencensusTraceSampled} %-5level - %msg";
+      "traceId=%X{traceId} spanId=%X{spanId} sampled=%X{traceSampled} %-5level - %msg";
 
   static final Tracestate EMPTY_TRACESTATE = Tracestate.builder().build();